contrib/azure-asr/requirements.txt
azure-cognitiveservices-speech
keyring
quipclient